+++ src/main/resources/templates/base.html
@@ -21,20 +21,8 @@
       {% block main_content %}{% endblock %}
     </div>
   </div>
-  <div class="konami-game">
-    <div class="konami" id="bird">
-    </div>
-    <div class="konami pipe" id="pipe-bottom">
-    </div>
-    <div class="konami pipe" id="pipe-top">
-    </div>
-    <div class="konami score">
-      <p id="score">0</p>
-    </div>
-  </div>
   {% block scripts %}
 	<script charset="utf-8" type="text/javascript" src="static/scripts/base.js"></script>
-	<script charset="utf-8" type="text/javascript" src="static/scripts/konami.js"></script>
   {% endblock %}
 </body>
