--- src/main/resources/templates/home.html
@@ -1,39 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Mentor Matching{% endblock %}
-
-{% block styles %}
-  {{ super() }}
-  {% set styles = [
-      "static/styles/home.css"
-    ] %}
-  {% for style in styles %}
-    <link rel="stylesheet" type="text/css" href={{ style }}>
-  {% endfor %}
-{% endblock %}
-
-{% block header %}
-  {% set nav_items = [
-      {
-        "baseURL": "/", "sectionID": "who-am-i", "iconLabel": "user", "text": "Who Am I?"
-      }
-    ] %}
-  {% include "navbar.html" %}
-  {% set title_image = "static/images/something.png" %}
-  {% set title = "Mentor Matching" %}
-  {% include "title-block.html" %}
-{% endblock %}
-
-{% block main_content %}
-
-{% endblock %}
-
-{% block scripts %}
-  {{ super() }}
-  {% set scripts = [
-      "static/scripts/home.js"
-    ]%}
-  {% for script in scripts %}
-    <script charset="utf-8" type="text/javascript" src={{ script }}></script>
-  {% endfor %}
-{% endblock %}
