+++ src/main/resources/templates/landing.html
@@ -28,13 +28,20 @@
       }
     ] %}
   {% include "navbar.html" %}
-  {% set title_image = "static/images/blank.png" %}
-  {% set title = "Mentor Matching" %}
-  {% include "title-block.html" %}
 {% endblock %}
 
 {% block main_content %}
-
+<div class="content-block" id="title-block">
+  <h1 class="page-title" id="title">Mentor Matching</h1>
+</div>
+<div class="content-block" id="join-mentor-matcher">
+  <button class="join-matcher-button pretty-button" id="join-as-mentee">Find a<br>mentor</button>
+  <button class="join-matcher-button pretty-button" id="join-as-mentor">Become a mentor</button>
+</div>
+<div class="content-block" id="member-sign-in">
+  <p class="short-tip" id="already-a-member">Already a member?</p>
+  <button class="pretty-button" id="login-button">Login</button>
+</div>
 {% endblock %}
 
 {% block scripts %}
