+++ src/main/webapp/static/styles/base.css
@@ -1,6 +1,13 @@
+:root {
+  --blue: #348edc;
+  --blue-lightened: #99c6ed;
+  --dark-blue: #4c81af;
+  --dark-blue-lightened: #a4c0d7;
+}
+
 html, body, .main-body {
   height: 100%;
-  height: -webkit-fill-available;
+  /* height: -webkit-fill-available; */
 }
 
 body {
@@ -20,6 +27,7 @@ body {
   align-items: center;
   min-width: 700px;
   width: 100%;
+  background-color: var(--dark-blue);
 }
 
 .main-body::after {
@@ -29,22 +37,19 @@ body {
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url("/static/images/blank.png");
+  /* background-image: url("/static/images/blank.png");
   background-repeat: no-repeat;
   background-size: cover;
   background-position: top;
   filter: brightness(1.2) blur(7px);
-  -webkit-filter: brightness(1.2) blur(7px);
+  -webkit-filter: brightness(1.2) blur(7px); */
 }
 
 .main-content {
-  max-width: 776px;
-  background-color: #ffffff99;
+  /* background-color: #ffffff99; */
   box-sizing: border-box;
-  margin: 1.5em 0;
-  padding: 1.5em;
-  border-radius: 1.5em;
   z-index: 1;
+  height: -webkit-fill-available;
 }
 
 .pretty-button {
@@ -79,7 +84,6 @@ body {
 .content-block {
   display: flex;
   flex-flow: column nowrap;
-  min-height: 30em;
   width: 100%;
   background-size: 100% 100%;
   background-repeat: no-repeat;
